Second Assignment for lesson 1
